Main Correções

+++ assets/js/calculadora.js
@@ -1,5 +1,5 @@
 function somaDigitos() {
-  const d1 = parseInt(document.getElementById("digito1").value);
-  const d2 = parseInt(document.getElementById("digito2").value);
+  const d1 = parseFloat(document.getElementById("digito1").value);
+  const d2 = parseFloat(document.getElementById("digito2").value);
   document.getElementById("result-sum").innerHTML = d1 + d2;
 }
